Fix bug with `VariableArray` by changing underlying implementation.
The previous implementation used `std::vector`, but this is unsafe for
`VariableArrayView`, because standard library vectors can have their
underlying data move around, invalidating pointers and references to
that data. Switching to a simple pointer to memory allocated by `new`
solves this problem, but requires additional constructors and a
destructor to be defined.